+++ src/main/java/tech/yump/vault/core/SealStatus.java
@@ -1,0 +1,6 @@
+package tech.yump.vault.core;
+
+public enum SealStatus {
+  SEALED,
+  UNSEALED
+}